Create dependency structure (#2)
This PR restructures the repo in order to handle multiple packages with dependency relations between them.

In detail:

Create a new libs folder and move the existing time lib under that folder
Add setup.py file for ebony-time, and move the dependency list into there, instead of requirements.txt files
Add a new dummy library dux under libs, which depends on time
Add compile_requirements.sh script which will produce a locked transitive requirements.txt file consisting of all the requirements of all the packages in the repo.
The idea here is that we can maintain the dependencies of each package independently, but for the purposes of defining the local environment, we just need to maintain a single requirements file which includes the requirements for every package in the repo. This also means that compiling the requirements should be pretty fast.

We might need to do something a bit smarter when it comes to thinking about deployment and testing that each package has declared all the dependencies it actually needs, but as a way to get the ball rolling I think this should do the trick.