Implement task library + refactor datagraph interface (#5)
This PR builds on top of the create-in-memory-data-graph branch, and adds a tasks library to libs/dux, as well as refactoring the existing data_graph library in a number of ways:

Simplifies the core interface of DataSetMetadata and DataSet - in particular decoupling them from the completion checking logic
Adds a new completion_checking module which defines an ICompletionChecker interface and two implementations (essentially regular + irregular)
Moves the logic to infer a completion checker from predecessors into the task library.
Adds calendars and time_of_day modules to the time lib, used by completion checking
Change semantics of TimeRange to be open at the end
Adds some utility methods to TimeRange
Rename from data_graph to datagraph because the compile requirements stuff wasn't handling the underscore correctly.
Additional details / motivations:

The core datagraph interface has been simplified so there is just IPersistenceEngine, DataSetMetadata, which contains precisely the information to uniquely identify a dataset (including the persistence engine), and DataSet, which is just a simple struct of metadata, data, and the declared_time_range.
Completion checking is factored out into a dedicated interface, with the paradigm being that you pass a DataSetMetadata into the is_complete method of a completion checker and get a response. This means that the completion checking logic can stay completely out of the way for cases where it is not relevant (e.g. static data, or local inspection).
The "glue" to associate a particular dataset with a particular completion checker has been moved into the task library - meaning in particular that that is where the "infer completion checking behaviour from predecessors" logic takes place. I think this is nice for a few reasons:
a) I think having the completion checker be part of the DatasetMetadata creates a bunch of annoying dilemmas. Do we persist it? If not, what does it get populated with in the datasets returned by a persistence engine? If so, what happens if we try to insert a new dataset that's the same as an existing one except for the completion checking? Do we update it, replace it, allow the two to coexist...? None of these problems seem particularly 'worth it', and so I think it's much nicer to just keep it decoupled.
b) We simply bypass completion checking entirely for static data, using a pure exists call instead.
c) Conceptually, the idea of having a fine-grained pre-specified definition of completeness is really only necessary for the purposes of resolving the dependencies in a computational graph, so having it be managed by the task framework just seems natural.
I think that maintaining the distinction between declared_time_range and data_time_range is a great idea, and to that end I've made declared_time_range a first-class member of DataSet, meaning in particular that every write method has to provide a value for it, and there's a dedicated read method to fetch it.